requirements.txt
fastapi==0.66.0
requests==2.26.0
uvicorn==0.14.0
python-dotenv==0.18.0
Babel==2.9.1
